+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+## (0.4.4) 22/10/24
+### Issue 
+- error en el movimiento de un peón
+### Changed
+- actualizacion de codigo de las piezas
 ## (0.4.3) 22/10/24
 ### Fixed
 - arreglo de actualizacion de la posicion, el metodo de traduccion de input estaba retornando de forma incorrecta las posiciones y eso me impedía realizar los movimientos correctos. :/